dcnn cleanup: moved dcnn code to dcnn/ dir
