+++ dcnn/Makefile
@@ -1,0 +1,8 @@
+INCLUDES=-I..
+
+OBJS=caffe.o dcnn.o dcnn_engine.o
+
+all: lib.a
+lib.a: $(OBJS)
+
+-include ../Makefile.lib
