+++ dcnn/caffe.cpp
