+++ pachi.c
@@ -15,7 +15,7 @@
 #include "engines/montecarlo.h"
 #include "engines/random.h"
 #include "engines/external.h"
-#include "engines/dcnn.h"
+#include "dcnn/dcnn_engine.h"
 #include "pattern/patternscan.h"
 #include "pattern/pattern_engine.h"
 #include "joseki/joseki_engine.h"
@@ -32,8 +32,8 @@
 #include "network.h"
 #include "uct/tree.h"
 #include "fifo.h"
-#include "dcnn.h"
-#include "caffe.h"
+#include "dcnn/dcnn.h"
+#include "dcnn/caffe.h"
 #include "pattern/pattern.h"
 #include "pattern/spatial.h"
 #include "pattern/prob.h"
